test(engine): extract interruption test to run without batch processing
The interruption of a terminating call activity by a timer boundary
event can only happen when the termination is not processed within a
batch. Here we disable batch processing by setting the limit to 1.
